refactor mock_decorator
- This will allow multiple functions to access the same code.
